pyproject.toml
[tool.poetry]
name = "grimmis"
version = "0.1.0"
description = "A stock trading app!"
authors = ["Aaron West <aphexlog@gmail.com>"]
license = "MIT"
# readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
polygon = "^1.2.0"
alpaca-trade-api = "^3.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venv = ".venv"
venvPath = "."

[tool.poetry.scripts]
grimmis = "grimmis.app:main"
